requirements.txt
websocket-client==1.8.0
python-dotenv==1.0.1
pyperclip==1.9.0
